# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: SCUD Backend API
  version: "1.0.0"
  description: |
    HTTP API системы управления доступом (SCUD).

    Контракт описывает три класса участников:

    - **Мобильное приложение пользователя** (Android) — авторизуется
      session_token'ом, выпускает ключи доступа, скачивает Bloom-фильтры для
      доставки на ридер, отправляет отчёты от ридера на сервер.
    - **Администратор / интеграция** — авторизуется API-ключом
      (`X-Api-Key`), управляет пользователями, группами, ридерами, permit'ами,
      выпущенными ключами, наблюдает доставку фильтров.
    - **Public** — health-check и пара unauthenticated auth-эндпоинтов
      (login / refresh).

    Ключевые понятия:

    - **Reader** — физический контроллер двери (ESP32). Имеет 16-байтный
      `reader_id` и Ed25519 публичный ключ.
    - **Reader group** — группа ридеров; пользователь видит только ридеры
      из доступной ему `user_group → reader_group` пары.
    - **Permit** — разрешение «пользователь P может пользоваться ридером R
      в окне [valid_from, valid_until], выпускать до n_parallel ключей с
      TTL ≤ max_token_ttl_seconds».
    - **IssuedKey** — выпущенный пользователем ключ (CARD на терминологию
      ридера). Подписан сервером, содержит `permit_id`, `device_id`, окно
      `[issued_at, expires_at]`. Отозванные ключи попадают в Bloom-фильтр
      ридера.
    - **TimeGrant** — подписанный сервером talisman, разрешающий телефону
...
      `"revoked"`).
servers:
...
security:
  - SessionToken: []
  - AdminApiKey: []
paths:
...
  /api/v1/app/auth/login:
    post:
      tags: [app-auth]
      summary: Логин по логину/паролю, получить session_token
      description: |
        Аутентифицирует пользователя по `login`/`password`. На успех создаёт
        запись Session (`session_token` живёт ~1 час, `refresh_token` —
        дольше) и возвращает оба токена + базовые данные пользователя.

        Если в запросе передан `device_info` — он пишется в Session как
        контекст устройства; сама регистрация ключа устройства (Ed25519
        pubkey) делается отдельно через `/auth/register-device`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthTokens"}
        "401":
          description: invalid_credentials
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorDetail"}}}
        "403":
          description: user_inactive
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorDetail"}}}
  /api/v1/app/auth/refresh:
    post:
      tags: [app-auth]
      summary: Обновить session_token по refresh_token
      description: |
        Меняет валидный `refresh_token` на свежую пару session+refresh.
        Старый refresh инвалидируется (rotation).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RefreshRequest"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthTokens"}
        "401":
          description: invalid_refresh_token
  /api/v1/app/auth/register-device:
    post:
      tags: [app-auth]
      summary: Привязать к сессии Ed25519-pubkey телефона
      description: |
        Создаёт (или переиспользует) запись Device для текущего пользователя,
        связывая её с активной Session. Без этого нельзя выпускать ключи и
        качать courier-доставки.

        Идемпотентно: если пользователь уже регистрировал тот же `phone_pubkey`,
        сервер возвращает существующий `device_id`. Если pubkey принадлежит
        другому пользователю — `409 pubkey_conflict`.
      security:
        - SessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RegisterDeviceRequest"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id: {type: string, format: uuid}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {description: "pubkey_conflict"}
        "422": {description: "invalid_base64"}
  /api/v1/app/auth/logout:
    post:
      tags: [app-auth]
      summary: Завершить текущую сессию
      description: |
        Помечает текущую Session как revoked. Refresh_token также становится
        невалидным.
      security:
        - SessionToken: []
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/Ok"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}

  # ===================================================================
  # APP / READERS
  # ===================================================================
...
  /api/v1/admin/users:
    post:
      tags: [admin-users]
      summary: Создать пользователя
      description: |
        Заводит запись User (логин/пароль/display_name/user_group). Пароль
        хешируется (argon2/bcrypt). На дубликат логина — **409 login_taken**.
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateUserRequest"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "409": {description: "login_taken"}
    get:
      tags: [admin-users]
      summary: Список пользователей
      description: Постраничный список всех пользователей (cursor-offset).
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminUser"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /api/v1/admin/users/{user_id}:
    get:
      tags: [admin-users]
      summary: Подробности пользователя
      security:
        - AdminApiKey: []
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/AdminUser"}}}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "user_not_found"}
    patch:
      tags: [admin-users]
      summary: Изменить пользователя
      description: |
        Меняет `display_name`, `user_group_id` или `is_active`. При
        `is_active=false` все активные сессии пользователя отзываются.
      security:
        - AdminApiKey: []
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PatchUserRequest"}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "user_not_found"}
...
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        `Authorization: Bearer <session_token>` — выдаётся `/auth/login` или
        `/auth/refresh`. Хранится на сервере в таблице sessions, валидируется
        каждый запрос (revoked_at IS NULL, expires_at > now).
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        `X-Api-Key: <plaintext>` — выдаётся `/admin/api-keys` один раз при
        создании. На сервере хранится только SHA256-хеш + `key_prefix`.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
      description: Offset для пагинации.
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Ok"}
    Unauthorized:
      description: missing_token / invalid_token / user_inactive
    AdminUnauthorized:
      description: missing_api_key / invalid_api_key / api_key_expired
  schemas:
    Ok:
      type: object
      properties:
        ok: {type: boolean, example: true}
    ErrorDetail:
      type: object
      description: |
        FastAPI стандартный формат — `{"detail": ...}`. detail может быть
        строкой (`"not_owner"`) или объектом (`{"error": "ttl_too_long",
        "max": 86400}`).
      properties:
        detail:
          oneOf:
            - {type: string}
            - {type: object}
...
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login: {type: string}
        password: {type: string}
        device_info:
          type: object
          properties:
            model: {type: string}
            os: {type: string}
            app_version: {type: string}
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: {type: string}
    RegisterDeviceRequest:
      type: object
      required: [phone_pubkey]
      properties:
        phone_pubkey:
          type: string
          description: Ed25519 pubkey телефона, base64 (32 байта).
        device_label:
          type: string
          nullable: true
    AuthTokens:
      type: object
      properties:
        session_token: {type: string}
        refresh_token: {type: string}
        user_id: {type: integer}
        user_group_id: {type: string, format: uuid}
        display_name: {type: string}

    # ----- READERS -----
...
    CreateUserRequest:
      type: object
      required: [login, password, display_name, user_group_id]
      properties:
        login: {type: string}
        password: {type: string}
        display_name: {type: string}
        user_group_id: {type: string, format: uuid}
    AdminUser:
      type: object
      properties:
        user_id: {type: integer}
        login: {type: string}
        display_name: {type: string}
        user_group_id: {type: string, format: uuid}
        is_active: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    PatchUserRequest:
      type: object
      properties:
        display_name: {type: string}
        user_group_id: {type: string, format: uuid}
        is_active: {type: boolean}

    # ----- PERMITS (APP) -----
...